# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Suraksha API",
        "description": "The `Suraksha API` will connect the `Sri Lanka Insurance Corporation\u2019s (SLIC) system` and `Student Information System (SIS) of the Ministry of Education, Sri Lanka` to make the claiming processes easier and accurate. The main Objective of the `Suraksha API` is to provide `SLIC` with necessary information from the `SIS` database and store the claim details in the `SIS` database.",
        "contact": {
            "email": "your-email@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/claim/add": {
            "post": {
                "tags": [
                    "claim"
                ],
                "summary": "Add a new claim to the system",
                "description": "",
                "operationId": "addClaim",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Claim object that needs to be added to the system",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Claim"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claim added successfully",
                        "schema": {
                            "$ref": "#/definitions/Claim"
                        }
                    },
                    "405": {
                        "description": "Failed to add the claim"
                    }
                }
            }
        },
        "/api/student{nsid}": {
            "get": {
                "tags": [
                    "student"
                ],
                "summary": "Search student by NSID",
                "description": "Returns a single student for the provided NSID",
                "operationId": "getStudentByNSID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "nsid",
                        "in": "path",
                        "description": "NSID of the student to return (ex: ABC-123-000)",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student found",
                        "schema": {
                            "$ref": "#/definitions/Student"
                        }
                    },
                    "404": {
                        "description": "Invalid NSID provided or Student not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "Claim": {
            "required": [
                "id",
                "incidentDate",
                "reqestedDate",
                "enteredDate",
                "type",
                "requestedAmount",
                "paidAmount",
                "payeeAddress"
            ],
            "properties": {
                "claim_id": {
                    "type": "string",
                    "example": "SLIC-00000"
                },
                "nsid": {
                    "type": "string",
                    "example": "ABC-123-000"
                },
                "student_name_with_initials": {
                    "type": "string",
                    "example": "J. Smith"
                },
                "student_full_name": {
                    "type": "string",
                    "example": "John Smith"
                },
                "student_date_of_birth": {
                    "type": "string",
                    "format": "date",
                    "example": "2000-01-15"
                },
                "mothers_name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "mothers_nic": {
                    "type": "string",
                    "example": "123456789V"
                },
                "fathers_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "fathers_nic": {
                    "type": "string",
                    "example": "123456789V"
                },
                "legal_guardians_name": {
                    "type": "string",
                    "example": "Legal Guardian's name - If any"
                },
                "incident_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2020-01-15"
                },
                "claim_requested_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2020-01-20"
                },
                "claim_entered_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2020-01-25"
                },
                "claim_type": {
                    "type": "string",
                    "example": "Residential"
                },
                "claim_description": {
                    "type": "string",
                    "example": "Claim description - If any"
                },
                "claim_amount_requested": {
                    "type": "integer",
                    "format": "int64",
                    "example": "100000"
                },
                "claim_amount_paid": {
                    "type": "integer",
                    "format": "int64",
                    "example": "100000"
                },
                "payment_receivers_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "payment_receivers_nic": {
                    "type": "string",
                    "example": "123456789V"
                },
                "payee_address": {
                    "type": "string",
                    "example": "Example Address, Area."
                },
                "claim_status": {
                    "type": "integer",
                    "format": "int8",
                    "example": "0"
                },
                "school_census_id": {
                    "type": "integer",
                    "format": "int32",
                    "example": "00000"
                },
                "school_name": {
                    "type": "string",
                    "example": "Example School"
                },
                "admission_id": {
                    "type": "integer",
                    "format": "int32",
                    "example": "00000"
                }
            },
            "type": "object"
        },
        "Student": {
            "required": [
                "nsid"
            ],
            "properties": {
                "nsid": {
                    "type": "string",
                    "example": "ABC-123-000"
                },
                "school": {
                    "type": "string",
                    "example": "Example School"
                },
                "schoolAddress": {
                    "type": "string",
                    "example": "Example School, Area."
                },
                "censusNo": {
                    "type": "integer",
                    "format": "int32",
                    "example": "00000"
                },
                "nameWithInitials": {
                    "type": "string",
                    "example": "J. Smith"
                },
                "fullName": {
                    "type": "string",
                    "example": "John Smith"
                },
                "admissionNo": {
                    "type": "integer",
                    "format": "int32",
                    "example": "00000"
                },
                "dob": {
                    "type": "string",
                    "format": "date",
                    "example": "2000-01-15"
                },
                "mothersName": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "mothersNIC": {
                    "type": "string",
                    "example": "123456789V"
                },
                "fathersName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "fathersNIC": {
                    "type": "string",
                    "example": "123456789V"
                },
                "guardiansName": {
                    "type": "string",
                    "example": "Legal Guardian's name - If any"
                }
            },
            "type": "object"
        }
    }
}